feat: 포인트 적립 API Service 기능 구현

+++ src/main/kotlin/com/kakaopay/honey/controller/dto/EarnPointDto.kt
@@ -3,7 +3,7 @@ package com.kakaopay.honey.controller.dto
 data class EarnPointRequestDto(
     val category: String,
     val membershipCode: String,
-    val point: Int,
+    val point: Long,
     val partnerName: String
 )
